front: persist scenario header
