+++ front/src/applications/operationalStudies/views/Scenario/ScenarioView.tsx
@@ -1,29 +1,22 @@
-import { useState } from 'react';
-
+import { usePersistScenarioHeader } from 'applications/operationalStudies/hooks/usePersistScenarioHeader';
 import useScenario from 'applications/operationalStudies/hooks/useScenario';
 import { ScenarioContextProvider } from 'applications/operationalStudies/hooks/useScenarioContext';
 import { RollingStockContextProvider } from 'common/RollingStockContext';
 import { SubCategoryContextProvider } from 'common/SubCategoryContext';
 
-import ScenarioHeader from './components/ScenarioHeader';
-import type { Board } from '../../types';
 import ScenarioContent from './components/ScenarioContent';
+import ScenarioHeader from './components/ScenarioHeader';
 
 const Scenario = () => {
   const { scenario } = useScenario();
 
-  const [activeBoards, setActiveBoards] = useState<Set<Board>>(
-    new Set<Board>(['trains', 'map', 'std', 'sdd', 'tables'])
-  );
-
-  const toggleBoard = (board: Board) => {
-    setActiveBoards((prev) => {
-      const newActiveBoards = new Set([...prev]);
-      if (newActiveBoards.has(board)) newActiveBoards.delete(board);
-      else newActiveBoards.add(board);
-      return newActiveBoards;
-    });
-  };
+  const { activeBoards, toggleBoard } = usePersistScenarioHeader(scenario?.id, [
+    'trains',
+    'map',
+    'std',
+    'sdd',
+    'tables',
+  ]);
 
   if (!scenario) return null;
